+++ data_pipeline/generic_web_api/module_constants.py
@@ -1,0 +1,9 @@
+
+
+# pylint: disable=too-few-public-methods
+class ModuleConstant:
+    DEFAULT_TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S%z"
+    DATA_IMPORT_TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S"
+    BQ_SCHEMA_FIELD_NAME_KEY = "name"
+    BQ_SCHEMA_SUBFIELD_KEY = "fields"
+    BQ_SCHEMA_FIELD_TYPE_KEY = "type"
